Moved TypeNameFormatter and more cleanup. All tests passed
